+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - ToolTip: Fixed calculation of position for selecting the opening side
+- DropZone: Prohibited deleting files in disabled mode 
 
 ## [0.32.0] - 2020-10-12